add encrypted file helper to enable using data for tests.  cleanup test names. add new tests.
