+++ UnitTestCommon/BrowseTreeParseTests.TestDifferentCaseFolderSort.verified.txt
@@ -1,0 +1,31 @@
+﻿
+Files: 0
+Children: 5
+a
+Files: 1
+Children: 0
+test
+B
+Files: 1
+Children: 2
+test
+B\anotherSubFolderOfB
+Files: 1
+Children: 0
+test
+B\subFolderOfB
+Files: 1
+Children: 0
+test
+b\anothersubofb
+Files: 1
+Children: 0
+test
+b\subofb
+Files: 1
+Children: 0
+test
+c
+Files: 1
+Children: 0
+test
