+++ UnitTestCommon/CachedParseResultsSerializerTests.RoundTrip_AllFieldsPreserved.received.txt
@@ -1,0 +1,131 @@
+﻿{
+  PresentableNameToFullFileInfo: {
+    key1: {
+      Item1: 100,
+      Item2: val1,
+      Item3: {
+        Item1: 1,
+        Item2: 2,
+        Item3: 3,
+        Item4: 4
+      },
+      Item4: true,
+      Item5: false
+    },
+    key2: {
+      Item1: 200,
+      Item2: val2,
+      Item3: {
+        Item1: 5,
+        Item2: 6,
+        Item3: 7,
+        Item4: 8
+      },
+      Item4: false,
+      Item5: true
+    }
+  },
+  DirectoryCount: 2,
+  BrowseResponse: {
+    Directories: [
+      {
+        Name: Music/Album1,
+        DecodedViaLatin1: false,
+        FileCount: 1,
+        Files: [
+          {
+            AttributeCount: 1,
+            Attributes: [
+              {
+                Value: 320
+              }
+            ],
+            BitRate: 320,
+            IsLatin1Decoded: false,
+            IsDirectoryLatin1Decoded: false,
+            Code: 1,
+            Extension: mp3,
+            Filename: song.mp3,
+            Size: 5000000
+          }
+        ],
+        OrderedFiles: [
+          {
+            AttributeCount: 1,
+            Attributes: [
+              {
+                Value: 320
+              }
+            ],
+            BitRate: 320,
+            IsLatin1Decoded: false,
+            IsDirectoryLatin1Decoded: false,
+            Code: 1,
+            Extension: mp3,
+            Filename: song.mp3,
+            Size: 5000000
+          }
+        ]
+      },
+      {
+        Name: Music/Album2,
+        DecodedViaLatin1: false,
+        FileCount: 1,
+        Files: [
+          {
+            IsLatin1Decoded: false,
+            IsDirectoryLatin1Decoded: false,
+            Code: 2,
+            Extension: flac,
+            Filename: track.flac,
+            Size: 20000000
+          }
+        ],
+        OrderedFiles: [
+          {
+            IsLatin1Decoded: false,
+            IsDirectoryLatin1Decoded: false,
+            Code: 2,
+            Extension: flac,
+            Filename: track.flac,
+            Size: 20000000
+          }
+        ]
+      }
+    ],
+    DirectoryCount: 2
+  },
+  BrowseResponseHiddenPortion: [
+    {
+      Name: Private/Stuff,
+      DecodedViaLatin1: false,
+      OrderedFiles: []
+    }
+  ],
+  PresentableDirectoryNameToDirectoryUriMappings: [
+    {
+      Item1: friendly1,
+      Item2: uri1
+    },
+    {
+      Item1: friendly2,
+      Item2: uri2
+    }
+  ],
+  SearchTermTokenToListOfFileKeys: {
+    token1: [
+      1,
+      2,
+      3
+    ],
+    token2: [
+      4,
+      5
+    ]
+  },
+  FileKeyToPresentableName: {
+    1: helper1,
+    2: helper2
+  },
+  NonHiddenFileCount: 42
+}
